requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
django-ninja==1.3.0
django-ninja-extra==0.22.0
django-ninja-jwt==5.3.5
dnspython==2.7.0
email_validator==2.2.0
fire==0.7.0
gunicorn==23.0.0
idna==3.10
injector==0.22.0
kombu==5.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
rav==0.0.9
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
